start adding an instructions class
